% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gee.fit}
\alias{gee.fit}
\title{Fit weighted GEE (point estimates only)}
\usage{
gee.fit(X, y, family, id, corstr, maxiter = 25, tol = 0.001,
  weights = NULL, scale, corr, calcVinv = F)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome vector}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
See \code{\link[stats]{family}} for details.}

\item{id}{a vector which identifies the clusters. The length of \code{id}
should be the same as the number of observations. Data are assumed to be
sorted so that observations on a cluster are contiguous rows for all
entities in the formula.}

\item{corstr}{character string specifying the correlation structure. Allowed
structures are: "independence", "exchangeable"}

\item{maxiter}{maximum Newton-Raphson updates of GEE parameters.}

\item{tol}{convergence tolerance for final GEE updates. Maximum absolute
difference between step (s-1) and (s) across all parameter values.}

\item{weights}{vector of weights corresponding to \eqn{y}}

\item{scale}{current value of scale parameter}

\item{corr}{current value of correlation parameter}

\item{calcVinv}{if T, calculates and outputs inverse of entire covariance
matrix V (note, this calculation could be very slow for correlation
parameters if cluster sizes are large, say >30 observations/cluster).}
}
\description{
This utility function fits a weighted GEE and returns parameter point
estimates, design matrix, and inverse of the working covariance matrix. We
solve for the regression parameters only; scale and correlation parameters
are assumed fixed at their given values. Note, this function can be used to
solve for second-order regression parameters such as scales or correlations;
in that case, scale and corr arguments of this function correspond to even
higher-order parameters (e.g. third and fourth moment quantities).
}
