% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gee_update}
\alias{gee_update}
\title{Update parameters of GEE using weighted Newton-Raphson update}
\usage{
gee_update(beta, D, V, W, y, mu, id)
}
\arguments{
\item{beta}{vector of parameters to update, e.g. \eqn{\beta} for marginal
means, \eqn{\gamma} for scales, \eqn{\alpha} for correlations}

\item{D}{matrix of partial derivatives}

\item{V}{working covariance matrix}

\item{W}{weight matrix}

\item{y}{vector of outcomes or residuals, e.g. \eqn{Y[k]}, \eqn{S[k]},
\eqn{Z[k]}}

\item{mu}{vector of marginal means, scales, or correlations, e.g.
\eqn{\mu[k]}, \eqn{\phi[k]}, \eqn{\rho[k]}}

\item{id}{vector of cluster identifiers}
}
\description{
Carries out one step of the usual GEE Newton-Raphson iterative fitting
procedure
}
