% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getVinv}
\alias{getVinv}
\title{Get Vinv}
\usage{
getVinv(x, family)
}
\arguments{
\item{x}{fitted geem object}

\item{family}{family argument from geem}
}
\description{
Obtains covariance matrix from fitted geem object.
}
