% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modNR}
\alias{modNR}
\title{Modified Newton-Raphson}
\usage{
modNR(ghat, r, maxit = 100, eps = 1e-07)
}
\arguments{
\item{ghat}{a \eqn{(J + 2L) x N} matrix, where \eqn{J} is the number of
propensity score models, \eqn{L} is the number of outcome regression models,
\eqn{N} is the total number of individuals in the data set (both observed and
unobserved). Row order must correspond to \code{r}}

\item{r}{an \eqn{N}-vector of 0s and 1s indicating which individuals had
their outcome observed (1) or missing (0). Order must correspond to rows in
\code{ghat}}

\item{maxit}{maximum number of iterations for algorithm; default
\code{maxit=100}}

\item{eps}{convergence criterion for algorithm; default \code{eps=10e-8}}
}
\description{
Carries out modified Newton-Raphson algorithm to solve for \eqn{\rho} for
MR-GEE.
}
\details{
This function inputs a \eqn{(J + 2L) x N} matrix \code{ghat} and outcome
observance \eqn{N}-vector \code{r}, performs modified Newton-Raphson, and
returns a \eqn{(J + 2L)}-vector \eqn{\hat{\rho}]}.
}
