% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mrgee.fit}
\alias{mrgee.fit}
\title{Fit Single Level of Multiply Robust GEE}
\usage{
mrgee.fit(formula, family, data, id, corstr = "independence",
  pmodels = NULL, bmodels = NULL, scale, corr, jointP = F,
  indiv_id = NULL, ...)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted.}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
See \code{\link[stats]{family}} for details.}

\item{data}{a data frame containing the variables in the model.}

\item{id}{a vector which identifies the clusters. The length of \code{id}
should be the same as the number of observations. Data are assumed to be
sorted so that observations on a cluster are contiguous rows for all
entities in the formula.}

\item{corstr}{character string specifying the correlation structure. Allowed
structures are: "independence", "exchangeable"}

\item{pmodels}{a list of propensity score models for \eqn{Pr(R=1|X, A)}, where
\eqn{R} indicates outcome observance, X are fully observed baseline
covariates, and A is the treatment indicator. List elements must be
objects of class "glm" from \code{\link[stats]{glm}}.}

\item{bmodels}{a list of outcome models for \eqn{E(Y|X, A)}, where \eqn{Y}
is the outcome, X are fully observed baseline covariates, and A is the
treatment indicator. List elements must be objects of class "gee" from
\code{\link[gee]{gee}}.}

\item{scale}{current value of scale parameter for mrgee.fit()}

\item{corr}{current value of correlation parameter for mrgee.fit()}

\item{jointP}{if jointP == T, we're requesting joint probabilities,
currently setup as just the products. Should be T if fitting MR-GEE for
correlation parameters.}

\item{indiv_id}{if provided, this corresponds to individual (rather than
joint) vector of ids. Should be provided if fitting MR-GEE for
correlation parameters.}

\item{...}{further arguments passed to \code{\link[mrgee]{modNR}}.}

\item{smodels}{a list of outcome models for \eqn{var(Y|X, A)}}

\item{cmodels}{a list of outcome models for \eqn{corr(Y, Y'|X, A)}}
}
\description{
A utility function used by \code{\link[mrgee]{mrgee}} to solve a single level
of MR-GEE.
}
\details{
This function solves a single weighted GEE corresponding to either the set of
marginal mean, scale, or correlation parameters in MR-GEE. This function is
called iteratively by \code{\link[mrgee]{mrgee}}. If used for correlation
parameter GEE, \code{jointP} should be TRUE and \code{indiv_id} should be
additionally specified.
}
\seealso{
\code{\link[mrgee]{mrgee}} for the general fitting function for
MR-GEE
}
