% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mrgee}
\alias{mrgee}
\title{Multiply Robust GEE}
\usage{
mrgee(formula, family = gaussian, data, id, corstr = "independence",
  scale.fix = F, pmodels = NULL, bmodels = NULL, smodels = NULL,
  cmodels = NULL, init.gamma = 1, init.alpha = 0.01, maxiter = 1,
  tol = 0.001, printIter = F)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted.}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
See \code{\link[stats]{family}} for details.}

\item{data}{a data frame containing the variables in the model.}

\item{id}{a vector which identifies the clusters. The length of \code{id}
should be the same as the number of observations. Data are assumed to be
sorted so that observations on a cluster are contiguous rows for all
entities in the formula.}

\item{corstr}{character string specifying the correlation structure. Allowed
structures are: "independence", "exchangeable"}

\item{scale.fix}{if T, then the scale parameter is fixed at init.gamma}

\item{pmodels}{a list of propensity score models for \eqn{Pr(R=1|X, A)}, where
\eqn{R} indicates outcome observance, X are fully observed baseline
covariates, and A is the treatment indicator. List elements must be
objects of class "glm" from \code{\link[stats]{glm}}.}

\item{bmodels}{a list of outcome models for \eqn{E(Y|X, A)}, where \eqn{Y}
is the outcome, X are fully observed baseline covariates, and A is the
treatment indicator. List elements must be objects of class "gee" from
\code{\link[gee]{gee}}.}

\item{smodels}{a list of outcome models for \eqn{var(Y|X, A)}}

\item{cmodels}{a list of outcome models for \eqn{corr(Y, Y'|X, A)}}

\item{init.gamma}{initial scale parameter value}

\item{init.alpha}{initial correlation parameter value}

\item{maxiter}{maximum Newton-Raphson updates of MR-GEE parameters (beta,
gamma, alpha).}

\item{tol}{convergence tolerance for final MR-GEE updates. Maximum absolute
difference between step (s-1) and (s) across all parameter updates (
i.e. all betas, gamma, alpha) is compared to tol.}

\item{printIter}{logical whether to print iteration information}
}
\description{
Fits a multiply robust generalized estimating equation (MR-GEE) using
empirical maximum likelihood to handle missing outcomes.
}
\details{
This function inputs a model formula, a list of propensity score models, and
lists of outcome regression models for the marginal means, scales, and
correlations to carry out MR-GEE. It outputs model fit information including
empirical MLEs for first- and second-order parameters (beta, gamma, alpha).
This function iteratively calls \code{\link[mrgee]{mrgee.fit}} to solve
MR-GEEs for each set of parameters (beta, gamma, alpha), corresponding to
parameters for the marginal mean, scales, and correlations, respectively.

Currently this function is limited to common scalar parameters for scale and
correlation. It also currently assumes propensity score model specified is
for individual probabilities and that joint probabilities are simply product
of individuals (i.e. assuming independence between missingness probabilities
).
}
\seealso{
\code{\link[mrgee]{mrgee.fit}} for the utility function used to
    solve each MR-GEE
}
