% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mrglm}
\alias{mrglm}
\title{Multiply Robust GLM}
\usage{
mrglm(formula, family = gaussian, data, pmodels, bmodels, ...)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. This argument is passed to
\code{\link[stats]{glm}}}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
This argument is passed to \code{\link[stats]{glm}}. See
\code{\link[stats]{family}} for details.}

\item{data}{a data frame containing the variables in the model. This argument
is passed to \code{\link[stats]{glm}}}

\item{pmodels}{a list of propensity score models for \eqn{Pr(R=1|X, S)}, where
\eqn{R} indicates outcome observance, X are covariates, and S are
auxiliary variables. List elements must be objects of class "glm" from
\code{\link[stats]{glm}}.}

\item{bmodels}{a list of outcome models for \eqn{E(Y|X, A)}, where \eqn{Y}
is the outcome, X are fully observed baseline covariates, and A is the
treatment indicator. List elements must be objects of class "glm" from
\code{\link[stats]{glm}}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Fits a multiply robust generalized linear model (MR-GLM) for independent data
using empirical maximum likelihood to handle missing outcomes. See Han
(2014).
}
\details{
This function inputs a model formula as well as lists of propensity score
models and outcome regression models to carry out multiply robust regression.
It outputs model fit information including empirical MLEs for the
coefficients.
}
