% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateV}
\alias{updateV}
\title{Update first-order working covariance}
\usage{
updateV(mu, gamma, alpha, id, corstr, VarFun)
}
\arguments{
\item{mu}{vector of marginal means}

\item{gamma}{scale parameters}

\item{alpha}{correlation parameters}

\item{id}{vector of cluster identifiers}

\item{corstr}{a character string specifying the correlation structure.
Allowed structures are: "independence" and "exchangeable"}

\item{VarFun}{function: the variance as a function of the mean}
}
\description{
Updates the entire working covariance matrix \eqn{V[1]} with most recent
second order parameter values, i.e. \eqn{(\gamma, \alpha)}
}
