% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mrgee}
\alias{mrgee}
\title{Multiply Robust GEE}
\usage{
mrgee(formula, family = gaussian, data, id, corstr = "independence",
  corr.mat = NULL, init.phi = 1, scale.fix = F, pmodels, bmodels,
  ...)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. This argument is passed to
\code{\link[geeM]{geem}}}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
This argument is passed to \code{\link[geeM]{geem}}. See
\code{\link[stats]{family}} for details.}

\item{data}{a data frame containing the variables in the model. This argument
is passed to \code{\link[geeM]{geem}}}

\item{id}{a vector which identifies the clusters. The length of \code{id}
should be the same as the number of observations. Data are assumed to be
sorted so that observations on a cluster are contiguous rows for all
entities in the formula.}

\item{corstr}{character string specifying the correlation structure. Allowed
structures are: "independence", "exchangeable", "ar1", "m-dependent",
"unstructured", "fixed", and "userdefined". Any unique substring may be
supplied. If "fixed" or "userdefined", then corr.mat must be specified. If
"m-dependent", then Mv is relevant.}

\item{pmodels}{a list of propensity score models for \eqn{Pr(R=1|X, A)}, where
\eqn{R} indicates outcome observance, X are fully observed baseline
covariates, and A is the treatment indicator. List elements must be
objects of class "glm" from \code{\link[stats]{glm}}.}

\item{bmodels}{a list of outcome models for \eqn{E(Y|X, A)}, where \eqn{Y}
is the outcome, X are fully observed baseline covariates, and A is the
treatment indicator. List elements must be objects of class "gee" from
\code{\link[gee]{gee}}.}

\item{...}{further arguments passed to \code{\link[mrgee]{modNR}}.}

\item{Mv}{for "m-dependent", the value for m.}
}
\description{
Fits a multiply robust generalized estimating equation (MR-GEE) using
empirical maximum likelihood to handle missing outcomes.
}
\details{
This function inputs a model formula as well as lists of propensity score
models and outcome regression models to carry out MR-GEE. It outputs model
fit information including empirical MLEs for the coefficients. The weighted
GEE is fit using \code{\link[geeM]{geem}}, which ensures consistency of this
weighted estimator, rather than the conventional  implementation in PROC
GENMOD for example.
}
